feat(reservation): 예약 생성 시 유효성 검사 및 과거 날짜 예약 방지 로직 추가

+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -2,16 +2,21 @@ package roomescape.service;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
+import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 import roomescape.dto.CreateReservationTimeRequest;
+import roomescape.exception.ReservationTimeInUseException;
+import roomescape.repository.ReservationDao;
 import roomescape.repository.ReservationTimeDao;
 
 @Service
 public class ReservationTimeService {
 
+    private final ReservationDao reservationDao;
     private final ReservationTimeDao reservationTimeDao;
 
-    public ReservationTimeService(ReservationTimeDao reservationTimeDao) {
+    public ReservationTimeService(ReservationDao reservationDao, ReservationTimeDao reservationTimeDao) {
+        this.reservationDao = reservationDao;
         this.reservationTimeDao = reservationTimeDao;
     }
 
@@ -25,6 +30,9 @@ public class ReservationTimeService {
     }
 
     public void deleteReservationTime(Long id) {
+        if(reservationDao.existsByTimeId(id)){
+            throw new ReservationTimeInUseException("해당 시간에 예약이 존재하여 삭제할 수 없습니다.");
+        }
         reservationTimeDao.deleteById(id);
     }
 }
